requirements.txt
django==3.1.7
mysqlclient==2.0.3
websocket-client==1.1.0
requests==2.25.1
django-redis==5.0.0
djangorestframework==3.12.4